refactor(mixin): remove private ProxyMixin
Inline proxy behavior directly into ObjectProxy.

Move shared proxy tracking helpers into runtime/lib/proxy-utils:
- ContentProxy
- contentFor
- customTagForProxy

Remove the private _ProxyMixin runtime export and delete the old
runtime/lib/mixins/-proxy module. Update internal imports, package
entrypoints, and tree-shakability expectations accordingly.

BREAKING CHANGE: private _ProxyMixin and runtime/lib/mixins/-proxy are removed.
ObjectProxy still provides content forwarding, unknownProperty,
setUnknownProperty, and isTruthy directly.
