+++ lib/index.js
@@ -40,7 +40,6 @@ const shim = addonV1Shim(path.join(__dirname, '..'), {
           './dist/dev/packages/@ember/-internals/routing/index.js',
           './dist/dev/packages/@ember/-internals/runtime/index.js',
           './dist/dev/packages/@ember/-internals/runtime/lib/ext/rsvp.js',
-          './dist/dev/packages/@ember/-internals/runtime/lib/mixins/-proxy.js',
           './dist/dev/packages/@ember/-internals/string/index.js',
           './dist/dev/packages/@ember/-internals/utility-types/index.js',
           './dist/dev/packages/@ember/-internals/utils/index.js',
