+++ tests/node-vitest/tree-shakability.test.js
@@ -113,7 +113,6 @@ it('[dev] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/-internals/routing/index.js",
       "ember-source/@ember/-internals/runtime/index.js",
       "ember-source/@ember/-internals/runtime/lib/ext/rsvp.js",
-      "ember-source/@ember/-internals/runtime/lib/mixins/-proxy.js",
       "ember-source/@ember/-internals/runtime/lib/mixins/action_handler.js",
       "ember-source/@ember/-internals/runtime/lib/mixins/container_proxy.js",
       "ember-source/@ember/-internals/runtime/lib/mixins/registry_proxy.js",
@@ -309,7 +308,6 @@ it('[prod] has expected tree-shakable entrypoints', async () => {
       "ember-source/@ember/-internals/routing/index.js",
       "ember-source/@ember/-internals/runtime/index.js",
       "ember-source/@ember/-internals/runtime/lib/ext/rsvp.js",
-      "ember-source/@ember/-internals/runtime/lib/mixins/-proxy.js",
       "ember-source/@ember/-internals/runtime/lib/mixins/action_handler.js",
       "ember-source/@ember/-internals/runtime/lib/mixins/container_proxy.js",
       "ember-source/@ember/-internals/runtime/lib/mixins/registry_proxy.js",
